Milestone 2.2 - Add scoring to quiz (#6)
Added ability to track attempt scores for each quiz attempt.  The main
screen lists the stats and a start button.  When the quiz is displayed
you can either quit and return to the main screen or take the quiz and
check the answers.  The score is added to the stats for the session.