Milestone 3.2 - Add quiz configuration options
Added ability to configure the number of questions for a quiz and
optionally set a time limit in minutes.  The questions are pulled
randomly from the question pool.  The number of questions can be from 1
to 15.  The timer can be set in minute increments up to 5.  A setting of
0 indicates no timer will be used.  Question inputs are disabled if
timer has elapsed.  Also refactored and cleaned up code.